create an initial test using Elasticsearch testcontainer.dotnet
