+++ src/ElasticLogger.Test/ElasticLogger.Test.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <!--<PackageReference Include="elasticsearch-inside" Version="6.6.1" />-->
+    <PackageReference Include="elasticsearch-inside" Version="6.6.1" />
     <PackageReference Include="Microsoft.AspNetCore.Hosting.Abstractions" Version="2.2.0" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.3" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.6.0" />
